# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Robot Fleet API
  description: Public API for QA engineers to practice API testing, automation,
    contract testing, performance testing, security testing and modern backend
    quality engineering.
  version: 1.0.0
  contact:
    name: Robot Fleet API
    url: https://github.com
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    def-0:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required:
        - page
        - limit
        - total
        - totalPages
      title: PaginationMeta
...
paths:
...
  /api/v1/robots/{id}/telemetry:
    get:
      summary: List telemetry records for a robot
      tags:
        - Telemetry
      parameters:
        - schema:
            minimum: 1
            default: 1
            type: integer
          in: query
          name: page
          required: false
        - schema:
            minimum: 1
            maximum: 100
            default: 20
            type: integer
          in: query
          name: limit
          required: false
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          format: uuid
                          type: string
                        robotId:
                          format: uuid
                          type: string
                        batteryLevel:
                          minimum: 0
                          maximum: 100
                          type: integer
                        temperature:
                          type: number
                        latitude:
                          type: number
                        longitude:
                          type: number
                        speed:
                          minimum: 0
                          type: number
                        recordedAt:
                          format: date-time
                          type: string
                      required:
                        - id
                        - robotId
                        - batteryLevel
                        - temperature
                        - latitude
                        - longitude
                        - speed
                        - recordedAt
                      title: Telemetry
                  pagination:
                    $ref: "#/components/schemas/def-0"
                required:
                  - data
                  - pagination
    post:
      summary: Record a telemetry reading for a robot (append-only)
      tags:
        - Telemetry
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                batteryLevel:
                  minimum: 0
                  maximum: 100
                  type: integer
                temperature:
                  minimum: -50
                  maximum: 150
                  type: number
                latitude:
                  minimum: -90
                  maximum: 90
                  type: number
                longitude:
                  minimum: -180
                  maximum: 180
                  type: number
                speed:
                  minimum: 0
                  maximum: 50
                  type: number
                recordedAt:
                  format: date-time
                  type: string
              required:
                - batteryLevel
                - temperature
                - latitude
                - longitude
                - speed
        required: true
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  robotId:
                    format: uuid
                    type: string
                  batteryLevel:
                    minimum: 0
                    maximum: 100
                    type: integer
                  temperature:
                    type: number
                  latitude:
                    type: number
                  longitude:
                    type: number
                  speed:
                    minimum: 0
                    type: number
                  recordedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - robotId
                  - batteryLevel
                  - temperature
                  - latitude
                  - longitude
                  - speed
                  - recordedAt
  /api/v1/telemetry/{id}:
    get:
      summary: Get a telemetry record by id
      tags:
        - Telemetry
      parameters:
        - schema:
            format: uuid
            type: string
          in: path
          name: id
          required: true
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    format: uuid
                    type: string
                  robotId:
                    format: uuid
                    type: string
                  batteryLevel:
                    minimum: 0
                    maximum: 100
                    type: integer
                  temperature:
                    type: number
                  latitude:
                    type: number
                  longitude:
                    type: number
                  speed:
                    minimum: 0
                    type: number
                  recordedAt:
                    format: date-time
                    type: string
                required:
                  - id
                  - robotId
                  - batteryLevel
                  - temperature
                  - latitude
                  - longitude
                  - speed
                  - recordedAt
...
servers:
  - url: /
    description: Current server
...